backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Environment and utilities
python-dotenv==1.0.0
pydantic==2.5.3
email-validator==2.1.1
pydantic-settings==2.1.0

# HTTP requests for Ollama
requests==2.31.0
httpx==0.26.0

# CORS
fastapi-cors==0.0.6

# Date/time
python-dateutil==2.8.2
